Use common superclass for unary CF and CI hybrids
Both AbstractUnaryHybridCF and AbstractUnaryHybridCI had exactly
the same code; let's consolidate into AbstractUnaryHybridC for both.
